AA: fix eventlog idempotence
This patch fixes the idempotence of eventlog. Before this, when AA
restarts and eventlog is activated, the originally recorded aael will be
truncated and the INIT event will be recorded repeatedly. This patch
will check whether there is an existing AAEL when AA is restarted. If
so, it will skip creating and recording the INIT event.

At the same time, a synchronization mechanism is used to ensure that
RTMR expansion will not occur repeatedly after AA abnormally interrupts
execution.

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>